starfish/requirements_dev.txt
ipython
ipdb
black
isort
flake8
mypy
python-decouple
django-extensions
pip-tools
pytest
coverage
pre-commit
factory_boy
sqlalchemy_mate>=1.4.28.3,<2.0.0.0
uszipcode
django-debug-toolbar
